Upgraded AppView object and modified Backbone.Router.gotoPage() method.
AppView object is now called PageManager and contains methods for
getting the current page uri hash fragment. This enables the router to
track the current page URI, which is required when dealing with modal
dialogs.

Router.gotoPage() is no longer used to manage loading of pages and
dialogs.  differences between the two are now handled directly by the
route callback functions.